requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyngrok==7.2.0

# Google OAuth & Authentication
google-auth==2.41.1
google-auth-oauthlib==1.2.2

# Cloud Storage & AI Generation
google-cloud-storage==3.4.1
google-cloud-aiplatform==1.78.0
pillow==11.3.0

# Image processing
aiofiles==24.1.0
